requirements.txt
networkx==2.6.3
nltk==3.6.7
numpy==1.22.2
python-dotenv==0.19.2
requests==2.27.1
requests_oauthlib==1.3.1
